backend/requirements.txt
asgiref==3.8.1
django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.7
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
